requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
torch==2.1.0
torchvision==0.16.0
facenet-pytorch==2.5.3
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
scikit-learn==1.3.2
